+++ ios/Exponent/Kernel/AppLoader/CachedResource/EXManifestResource.m
@@ -8,10 +8,11 @@
 #import "EXKernelLinkingManager.h"
 #import "EXKernelUtil.h"
 #import "EXVersions.h"
-#import <EXManifests/EXManifestsManifestFactory.h>
 
 #import <React/RCTConvert.h>
-#import <EXUpdates/EXUpdatesUpdate.h>
+
+@import EXManifests;
+@import EXUpdates;
 
 NSString * const kEXPublicKeyUrl = @"https://exp.host/--/manifest-public-key";
 NSString * const EXRuntimeErrorDomain = @"incompatible-runtime";