+++ packages/@expo/cli/src/export/exportAsync.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 
 import * as Log from '../log';
+import { FileNotifier } from '../utils/FileNotifier';
 import { ensureDirectoryAsync, removeAsync } from '../utils/dir';
 import { exportAppAsync } from './exportApp';
 import { Options } from './resolveOptions';
@@ -15,6 +16,9 @@ export async function exportAsync(projectRoot: string, options: Options) {
 
   // Export the app
   await exportAppAsync(projectRoot, options);
+
+  // Stop any file watchers to prevent the CLI from hanging.
+  FileNotifier.stopAll();
 
   // Final notes
   Log.log(`Export was successful. Your exported files can be found in ${options.outputDir}`);