+++ packages/expo-updates/e2e/fixtures/Updates-assets.e2e.ts
@@ -7,7 +7,7 @@ import Update from './utils/update';
 
 const projectRoot = process.env.PROJECT_ROOT || process.cwd();
 const platform = (process.env.DETOX_CONFIGURATION || 'ios.release').split('.')[0];
-
+const protocolVersion = platform === 'android' ? 1 : 0;
 const TIMEOUT_BIAS = process.env.CI ? 10 : 1;
 
 /**
@@ -34,7 +34,7 @@ describe('Asset deletion recovery', () => {
      * DatabaseLauncher should copy all the missing assets and run the update as normal.
      */
     jest.setTimeout(300000 * TIMEOUT_BIAS);
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
 
     /**
      * Install the app and immediately send it a message to clear internal storage. Verify storage
@@ -115,7 +115,7 @@ describe('Asset deletion recovery', () => {
      * the missing assets and run the update as normal.
      */
     jest.setTimeout(300000 * TIMEOUT_BIAS);
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
 
     /**
      * Install the app and immediately send it a message to clear internal storage. Verify storage
@@ -205,7 +205,7 @@ describe('Asset deletion recovery', () => {
 
     const bundledAssets = Update.findAssets(projectRoot, platform);
     const assets = await Promise.all(
-      bundledAssets.map(async (asset: { path: string; ext: string; }) => {
+      bundledAssets.map(async (asset: { path: string; ext: string }) => {
         const filename = path.basename(asset.path);
         const mimeType = asset.ext === 'ttf' ? 'font/ttf' : 'image/png';
         const key = filename.replace('asset_', '').replace(/\.[^/.]+$/, '');
@@ -230,7 +230,7 @@ describe('Asset deletion recovery', () => {
     /**
      * Install the app and launch it so that it downloads the new update we're hosting
      */
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({ newInstance: true });