+++ packages/expo-updates/e2e/fixtures/Updates-basic.e2e.ts
@@ -3,12 +3,11 @@ import path from 'path';
 import { setTimeout } from 'timers/promises';
 
 import Server from './utils/server';
-
 import Update from './utils/update';
 
 const projectRoot = process.env.PROJECT_ROOT || process.cwd();
 const platform = (process.env.DETOX_CONFIGURATION || 'ios.release').split('.')[0];
-
+const protocolVersion = platform === 'android' ? 1 : 0;
 const TIMEOUT_BIAS = process.env.CI ? 10 : 1;
 
 describe('', () => {
@@ -19,7 +18,7 @@ describe('', () => {
 
   it('starts app, stops, and starts again', async () => {
     jest.setTimeout(300000 * TIMEOUT_BIAS);
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await device.installApp();
     await device.launchApp({
       newInstance: true,
@@ -75,7 +74,7 @@ describe('', () => {
       []
     );
 
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({
@@ -118,7 +117,7 @@ describe('', () => {
       []
     );
 
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({
@@ -179,7 +178,7 @@ describe('', () => {
       assets
     );
 
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({
@@ -262,7 +261,7 @@ describe('', () => {
       assets
     );
 
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({
@@ -300,7 +299,7 @@ describe('', () => {
       []
     );
 
-    Server.start(Update.serverPort);
+    Server.start(Update.serverPort, protocolVersion);
     await Server.serveSignedManifest(manifest, projectRoot);
     await device.installApp();
     await device.launchApp({
@@ -319,5 +318,85 @@ describe('', () => {
     await device.launchApp();
     const secondMessage = await Server.waitForRequest(10000 * TIMEOUT_BIAS);
     expect(secondMessage).toBe('test');
+  });
+
+  it('supports rollbacks', async () => {
+    if (platform === 'ios') {
+      console.warn('Rollbacks not yet implemented on iOS: exiting.');
+      return;
+    }
+    jest.setTimeout(300000 * TIMEOUT_BIAS);
+    const bundleFilename = 'bundle1.js';
+    const newNotifyString = 'test-update-3';
+    const hash = await Update.copyBundleToStaticFolder(
+      projectRoot,
+      bundleFilename,
+      newNotifyString,
+      platform
+    );
+    const manifest = Update.getUpdateManifestForBundleFilename(
+      new Date(),
+      hash,
+      'test-update-3-key',
+      bundleFilename,
+      []
+    );
+
+    Server.start(Update.serverPort, protocolVersion);
+    await Server.serveSignedManifest(manifest, projectRoot);
+    await device.installApp();
+    await device.launchApp({
+      newInstance: true,
+    });
+    const firstRequest = await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
+    const message = await Server.waitForRequest(10000 * TIMEOUT_BIAS);
+    expect(message).toBe('test');
+
+    // give the app time to load the new update in the background
+    await setTimeout(2000 * TIMEOUT_BIAS);
+    expect(Server.consumeRequestedStaticFiles().length).toBe(1);
+
+    // restart the app so it will launch the new update
+    await device.terminateApp();
+    await device.launchApp();
+    const secondRequest = await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
+    const updatedMessage = await Server.waitForRequest(10000 * TIMEOUT_BIAS);
+    expect(updatedMessage).toBe(newNotifyString);
+
+    // serve a rollback now
+    const rollbackDirective = Update.getRollbackDirective(new Date());
+    await Server.serveSignedDirective(rollbackDirective, projectRoot);
+
+    // restart the app so it will fetch the rollback
+    await device.terminateApp();
+    await device.launchApp();
+    const thirdRequest = await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
+    await Server.waitForRequest(10000 * TIMEOUT_BIAS);
+
+    // Restart the app so it will launch the rollback
+    await device.terminateApp();
+    await device.launchApp();
+    const fourthRequest = await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
+    const rolledBackMessage = await Server.waitForRequest(10000 * TIMEOUT_BIAS);
+    expect(rolledBackMessage).toBe('test');
+    expect(secondRequest.headers['expo-embedded-update-id']).toBeDefined();
+    expect(secondRequest.headers['expo-embedded-update-id']).toEqual(
+      firstRequest.headers['expo-embedded-update-id']
+    );
+    expect(thirdRequest.headers['expo-embedded-update-id']).toEqual(
+      firstRequest.headers['expo-embedded-update-id']
+    );
+    expect(fourthRequest.headers['expo-embedded-update-id']).toEqual(
+      firstRequest.headers['expo-embedded-update-id']
+    );
+
+    expect(firstRequest.headers['expo-current-update-id']).toEqual(
+      firstRequest.headers['expo-embedded-update-id']
+    );
+    expect(secondRequest.headers['expo-current-update-id']).toEqual(manifest.id);
+    expect(thirdRequest.headers['expo-current-update-id']).toEqual(manifest.id);
+    expect(fourthRequest.headers['expo-current-update-id']).toEqual(
+      firstRequest.headers['expo-embedded-update-id']
+    );
   });
 });